[update] 관리자 상품 관리 수정
- 삭제시 confirm 창 뜨도록 수정

+++ frontend/src/pages/admin/CreateProduct/components/CreateCategory/index.tsx
@@ -29,7 +29,9 @@ const CreateCatgegory = () => {
   };
 
   const deleteCategoryHandler = () => {
-    dispatch(deleteCategoryThunk(selectedCategory));
+    if (window.confirm('정말 삭제하시겠습니까?')) {
+      dispatch(deleteCategoryThunk(selectedCategory));
+    }
   };
 
   return (
